abstract everything cling behind upnpclient
